feat: new auth adapter (continuation)
